# Conflicts:
#	SaintCoinach/Definitions/EventIconPriorityPair.json

+++ SaintCoinach/IO/IndexFile.cs
@@ -10,7 +10,7 @@ namespace SaintCoinach.IO {
         public PackIdentifier PackId { get; private set; }
         public uint FileKey { get; private set; }
         public uint DirectoryKey { get; private set; }
-        public int Offset { get; private set; }
+        public uint Offset { get; private set; }
 
         /// <summary>
         ///     In which .dat* file the data is located.
@@ -28,7 +28,7 @@ namespace SaintCoinach.IO {
 
             var baseOffset = reader.ReadInt32();
             DatFile = (byte)((baseOffset & 0x000F) / 2);
-            Offset = (baseOffset - (baseOffset & 0x000F)) * 0x08;
+            Offset = (uint)(baseOffset - (baseOffset & 0x000F)) * 0x08;
 
             reader.ReadInt32(); // Zero
         }
